wininet: Convert request to UTF8 and % encode
pulling in the latest wininet changes from Wine to fix the ssl over proxy issue rolled out
my previous changes for handling international characters.  with this commit, i re-merge
those fixes.  the only change is to check for NULL on return from heap_realloc

looking back on this code, i think the correct fix for the original UTF8/% is to add support
for URL_ESCAPE_AS_UTF8 to UrlEscapeW, and use it in HttpOpenRequestW (and probably HTTP_HandleRedirect)
but that would be new development and the current fixes have been known to work for years