ntdll: llheap: Reuse orphaned buffers before allocating new ones.
threads orphan their buffers when they exit.  these are currently all adopted
by the next thread that allocates memory.  but this can result in wasted memory
because any holes in any but the first buffer with room are not immediately
filled.  and because they are now owned by a new thread, another thread cannot
use them

this change changes the model so that it does not depend on the new owner
using all of the memory in the buffers it adopts.  only the adopts the orphans
that satisfy the memory request.