ntdll: Fix race condition with fd_cache.
Signed-off-by: Daniel Lehman <dlehman25@gmail.com>

There's a race condition with DuplicateHandle and the fd_cache
It happens in the following situation:
- Thread 1 calls DuplicateHandle(CLOSE_SOURCE), which is 3 operations
    - req_dup_handle
    - remove_fd_from_cache
    - close
- Thread 2 calls CreateFile, then ReadFile (which calls server_get_unix_fd)
    - CreateFile
        - req_create_file
    - ReadFile
        - server_get_unix_fd
            - get_cached_fd (before lock)
            - if uncached:
                - get_cached_fd (inside lock)
                - req_get_handle_fd
                - add_fd_to_cache
        - read/pread

The problem is when a thread caches a handle in the fd_cache, then duplicates and closes it.
The closing of the handle is done on the wineserver during duplication, before it is removed
from the fd_cache on the client side.  This allows another thread opening a file to use the
just-closed handle for another file.  Because the fd_cache entry for that handle hasn't been
invalidated by the first thread, another thread can see the old mapping.  This can lead
to accessing the wrong fd or seeing an invalid fd if the first thread calls close().

T1                                  T2                          fd_cache    Notes
---------------------               --------------------------- ----------- ------------------------
(T1 opened 'foo.txt' as 0x100)      -                           0x100 -> 4  T1 has mapped 0x100 as fd 4 in fd_cache
req_dup_handle(0x100, CLOSE_SOURCE) -                           0x100 -> 4  T1 duplicate handle closing source
-                                   -                           0x100 -> 4  wineserver closes 0x100, now available
-                                   req_create_file(bar.txt)    0x100 -> 4  T2 opens 'bar.txt', gets 0x100, but fd_cache still has 0x100 -> 4
-                                   get_cached_fd(0x100) -> 4   0x100 -> 4  T2 sees existing 0x100 in fd_cache, gets fd 4
remove_fd_from_cache(0x100)         -                           0x100 -> X  T1 removes handle from cache _after_ T2 fetched it
-                                   pread(4 -> 'foo.txt)        0x100 -> X  T2 reads from fd 4 -> 'foo.txt' instead of 'bar.txt'
close(4)                            -                                       T1 finally closes fd 4.  this could also happen
                                                                            before T2's read and lead to EBADF

The fix here is to remove the fd from the cache first, as in NtClose.  If the file is really being
closed, then there's no need for it in the cache anyway.  If it really wasn't closed, it'll just be
re-cached again when used.  It is expected this is a rare case, so should not affect performance