dsound: Flush denormals in the mixing thread.
I suspect denormals might be the cause of the unexplained slowness
mentioned in a few of the later comments on bug 30639. Denormals can
be much slower than normal floats, although the penalty is generally
significant only on older CPUs. 64-bit processes default to flushing,
so this only really matters for 32-bit.